requirements.txt
asgiref==3.5.2
awsebcli==3.20.3
botocore==1.23.54
cement==2.8.2
certifi==2022.9.14
charset-normalizer==2.0.12
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
future==0.16.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.1
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
PyJWT==2.5.0
pyparsing==3.0.9
pypiwin32==223
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
pywin32==304
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.1.9
